Time: 4 ms (83.71%), Space: 38.1 MB (25.9%) - LeetHub

+++ 1976-number-of-ways-to-arrive-at-destination/1976-number-of-ways-to-arrive-at-destination.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int countPaths(int n, vector<vector<int>>& roads) {
-        vector<vector<pair<int,int>>> adj(n);
+        vector<vector<pair<int,long long>>> adj(n);
         long long  mod = (int)(1e9+7);
         priority_queue<pair<long long,int>, vector<pair<long long,int>>, greater<pair<long long,int>>> q;
         vector<long long> dist (n,LLONG_MAX);
